refactor : Seperate InputViewTest class to ValidatorTest class and InputViewTest class
