+++ src/test/java/baseball/view/InputViewTest.java
@@ -1,7 +1,6 @@
 package baseball.view;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.io.ByteArrayInputStream;
 import org.junit.jupiter.api.Test;
@@ -24,59 +23,5 @@ public class InputViewTest {
         userInput("1");
         int userNumbers = InputView.getRestart();
         assertThat(userNumbers).isEqualTo(1);
-    }
-
-    @Test
-    void 문자_포함_예외_테스트() {
-        userInput("a34");
-
-        Exception e = assertThrows(IllegalArgumentException.class, InputView::getUserNumbers);
-        assertThat(e.getClass()).isEqualTo((IllegalArgumentException.class));
-        assertThat(e.getMessage()).isEqualTo("숫자를 입력해주세요.");
-    }
-
-    @Test
-    void 입력_길이_미만_예외_테스트() {
-        userInput("12");
-
-        Exception e = assertThrows(IllegalArgumentException.class, InputView::getUserNumbers);
-        assertThat(e.getClass()).isEqualTo(IllegalArgumentException.class);
-        assertThat(e.getMessage()).isEqualTo("3자리의 수를 입력해주세요.");
-    }
-
-    @Test
-    void 입력_길이_초과_예외_테스트() {
-        userInput("1234");
-
-        Exception e = assertThrows(IllegalArgumentException.class, InputView::getUserNumbers);
-        assertThat(e.getClass()).isEqualTo(IllegalArgumentException.class);
-        assertThat(e.getMessage()).isEqualTo("3자리의 수를 입력해주세요.");
-    }
-
-    @Test
-    void 입력_숫자_중복_예외_테스트() {
-        userInput("223");
-
-        Exception e = assertThrows(IllegalArgumentException.class, InputView::getUserNumbers);
-        assertThat(e.getClass()).isEqualTo(IllegalArgumentException.class);
-        assertThat(e.getMessage()).isEqualTo("서로 다른 3자리의 수를 입력해주세요.");
-    }
-
-    @Test
-    void 입력_0값_예외_테스트() {
-        userInput("012");
-
-        Exception e = assertThrows(IllegalArgumentException.class, InputView::getUserNumbers);
-        assertThat(e.getClass()).isEqualTo(IllegalArgumentException.class);
-        assertThat(e.getMessage()).isEqualTo("0이 포함되지 않는 수를 입력해주세요.");
-    }
-
-    @Test
-    void 게임_종료_입력_예외_테스트() {
-        userInput("0");
-
-        Exception e = assertThrows(IllegalArgumentException.class, InputView::getRestart);
-        assertThat(e.getClass()).isEqualTo(IllegalArgumentException.class);
-        assertThat(e.getMessage()).isEqualTo("1, 2 중 하나를 입력해주세요.");
     }
 }
