+++ src/test/java/baseball/view/OutputViewTest.java
@@ -17,20 +17,22 @@ public class OutputViewTest {
 
     @Test
     void 연산_결과_출력_테스트() {
+        String newline = System.lineSeparator();
+
         OutputStream os = computerResult();
         OutputView.printBaseballResult(1, 0);
-        assertThat(os.toString()).isEqualTo("1볼 \n");
+        assertThat(os.toString()).isEqualTo("1볼 " + newline);
 
         os = computerResult();
         OutputView.printBaseballResult(0, 1);
-        assertThat(os.toString()).isEqualTo("1스트라이크\n");
+        assertThat(os.toString()).isEqualTo("1스트라이크" + newline);
 
         os = computerResult();
         OutputView.printBaseballResult(1, 1);
-        assertThat(os.toString()).isEqualTo("1볼 1스트라이크\n");
+        assertThat(os.toString()).isEqualTo("1볼 1스트라이크" + newline);
 
         os = computerResult();
         OutputView.printBaseballResult(0, 0);
-        assertThat(os.toString()).isEqualTo("낫싱\n");
+        assertThat(os.toString()).isEqualTo("낫싱" + newline);
     }
 }
